+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix `BalancesSubscription` sending null for snapshot by default (skip_serializing_none)
 - Change `checksum` fields for L2 and L3 books to `u32` to match crc32fast crate impl
+- Derive Clone for `Message<T>` and all `*Subscription` types
 
 ### v0.1.0